+++ android/rctmln/src/main/java/com/maplibre/rctmln/components/location/LocationComponentManager.java
@@ -67,6 +67,15 @@ public class LocationComponentManager {
         }
     }
 
+    public void setPreferredFramesPerSecond(int preferredFramesPerSecond) {
+       if(preferredFramesPerSecond <= 0) {
+            return;
+       }
+
+        mLocationComponent.setMaxAnimationFps(preferredFramesPerSecond);
+    }
+
+
     public void addOnCameraTrackingChangedListener(OnCameraTrackingChangedListener onCameraTrackingChangedListener) {
         mLocationComponent.addOnCameraTrackingChangedListener(onCameraTrackingChangedListener);
     }