+++ plugin/src/__tests__/fixtures/cocoapodFiles.ts
@@ -2,7 +2,7 @@ export const reactNativeTemplatePodfile = `
 require_relative '../node_modules/react-native/scripts/react_native_pods'
 require_relative '../node_modules/@react-native-community/cli-platform-ios/native_modules'
 
-platform :ios, '11.0'
+platform :ios, '12.0'
 
 target 'HelloWorld' do
   config = use_native_modules!
@@ -35,7 +35,7 @@ require_relative '../node_modules/react-native/scripts/react_native_pods'
 require_relative '../node_modules/react-native-unimodules/cocoapods.rb'
 require_relative '../node_modules/@react-native-community/cli-platform-ios/native_modules'
 
-platform :ios, '11.0'
+platform :ios, '12.0'
 
 target 'HelloWorld' do
   use_unimodules!
@@ -59,7 +59,7 @@ require_relative '../node_modules/react-native/scripts/react_native_pods'
 require_relative '../node_modules/react-native-unimodules/cocoapods.rb'
 require_relative '../node_modules/@react-native-community/cli-platform-ios/native_modules'
 
-platform :ios, '11.0'
+platform :ios, '12.0'
 
 target 'HelloWorld' do
   use_unimodules!
@@ -87,7 +87,7 @@ require_relative '../node_modules/react-native/scripts/react_native_pods'
 require_relative '../node_modules/react-native-unimodules/cocoapods.rb'
 require_relative '../node_modules/@react-native-community/cli-platform-ios/native_modules'
 
-platform :ios, '11.0'
+platform :ios, '12.0'
 
 target 'HelloWorld' do
   use_unimodules!
@@ -128,7 +128,7 @@ require_relative '../node_modules/react-native/scripts/react_native_pods'
 require_relative '../node_modules/react-native-unimodules/cocoapods.rb'
 require_relative '../node_modules/@react-native-community/cli-platform-ios/native_modules'
 
-platform :ios, '11.0'
+platform :ios, '12.0'
 
 target 'HelloWorld' do
   use_unimodules!
@@ -166,7 +166,7 @@ end
 `;
 
 export const blankTemplatePodfile = `
-platform :ios, '11.0'
+platform :ios, '12.0'
 
 target 'HelloWorld' do
 end