+++ CHANGELOG.md
@@ -6,6 +6,8 @@ PR Title ([#123](link to my pr))
 
 ```
 
+Chore: [bump maplibre native ios to 6.5.4](https://github.com/maplibre/maplibre-react-native/pull/437)
+
 ## 10.0.0-alpha.10
 
 Fix: [move @types/ packages to deps and remove assets.d.ts #423](https://github.com/maplibre/maplibre-react-native/pull/423)